Feature/#136 sign up complete save transaction record event

+++ porko-service/src/main/java/io/porko/history/domain/HistoryCategory.java
@@ -8,20 +8,17 @@ import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.AccessType;
 import org.springframework.data.annotation.AccessType.Type;
 
-import java.math.BigInteger;
-
 @Getter
 @Embeddable
 @AccessType(Type.FIELD)
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class HistoryCategory {
-
-    private BigInteger imageUrlTypeNo;
-    private String type; // 수입인지 지출인지를 나타냄.
+    private int imageUrlTypeNo;
+    private String type;
     private String name;
 
-    public static HistoryCategory of(BigInteger imageUrlTypeNo, String type, String name) {
+    public static HistoryCategory of(int imageUrlTypeNo, String type, String name) {
         return new HistoryCategory(imageUrlTypeNo, type, name);
     }
 }
